products: move Unit, Amount, Product to products app
