+++ src/apps/recipes/models.py
@@ -1,27 +1,6 @@
 from django.db import models
 
-
-class Unit(models.Model):
-    name = models.CharField(max_length=50, unique=True)
-
-    def __str__(self):
-        return '%s' % (self.name)
-
-
-class Amount(models.Model):
-    # TODO: maybe DecimalField would be better?
-    weight = models.FloatField()
-    unit = models.ForeignKey(Unit, on_delete=models.CASCADE)
-
-    def __str__(self):
-        return f'{round(self.weight, 2)} {self.unit}'
-
-
-class Product(models.Model):
-    name = models.CharField(max_length=300, unique=True)
-
-    def __str__(self):
-        return '%s' % (self.name)
+from apps.products.models import Amount, Product
 
 
 class Ingredient(models.Model):
